Ensure sync wait is returned as an int to fix multisite bug

+++ sync/class.jetpack-sync-sender.php
@@ -213,7 +213,7 @@ class Jetpack_Sync_Sender {
 	}
 
 	function get_sync_wait_time() {
-		return get_option( self::SYNC_THROTTLE_OPTION_NAME );
+		return (int) get_option( self::SYNC_THROTTLE_OPTION_NAME );
 	}
 
 	private function get_last_sync_time() {
